Update initConfigParams.m
Attempt at fixing an error caused when the soil moisture list is represented by a row vector instead of a column vector.

This seems to be specific to soil moisture and it can be seen in files such as updateGndDynParams as shown:

VSM_cm3cm3 = VSM_list_cm3cm3( sim_counter, : )';
RMSH_cm = RMSH_list_cm( sim_counter );

In this case, RMSH will work fine as the indexing works regardless of it being represented as row or column vector. VSM, however, will be incorrect.

This might cause errors in other parts of the code, so we'll need to be careful.